+++ CHANGELOG.md
@@ -1,6 +1,14 @@
+## 1.0.3
+
+- renamed `InputCalendar` to `InputDatePicker` because we are working
+ on a calendar widget containing events. Sorry for any inconvenience.
+- added first animated gif. Shows the date picker.
+
 ## 1.0.2
 
 - `InputSpinner` now accepts a generic of type `int` or `double`.
+- Date utility methods (e.g. `weekOfYear` or `julianDay`) changed
+ to Dart extensions. Can now be used on all `DateTime` objects.
 - Bugfix: closing `InputCalendar` with \[X\] did set value to null
 
 ## 1.0.1
@@ -31,4 +39,3 @@
 ## 0.1.0
 
 - Initial version
-